refactor: LottoScore 삭제
