+++ src/test/kotlin/lotto/domain/LottoResultsTest.kt
@@ -1,0 +1,30 @@
+package lotto.domain
+
+import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.params.ParameterizedTest
+import org.junit.jupiter.params.provider.CsvSource
+import java.math.BigDecimal
+import java.math.RoundingMode
+
+internal class LottoResultsTest {
+
+    @ParameterizedTest
+    @CsvSource(
+        value = [
+            "5000, 12",
+            "60000, 1",
+            "60500, 1",
+            "120000, 0.5",
+            "210000, 0.285"
+        ]
+    )
+    internal fun `당첨 결과에 따라 수익률을 반환한다`(price: Int, expected: BigDecimal) {
+        // 총 상금: 60,000
+        val lottoResult = listOf(
+            LottoResult(LottoPrize.FOURTH, 1),
+            LottoResult(LottoPrize.FIFTH, 2)
+        ).toLottoResults()
+        val result = lottoResult.rateOfResult(LottoPrice(price))
+        assertThat(result).isEqualTo(expected.setScale(2, RoundingMode.HALF_UP))
+    }
+}
