+++ src/test/kotlin/lotto/domain/LottoTicketsTest.kt
@@ -3,12 +3,8 @@ package lotto.domain
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
-import org.junit.jupiter.params.ParameterizedTest
-import org.junit.jupiter.params.provider.CsvSource
-import java.math.BigDecimal
-import java.math.RoundingMode
 
-internal class LottoScoreTest {
+internal class LottoTicketsTest {
     @Test
     internal fun `당첨 번호와 나의 로또 번호를 비교하여 결과를 반환한다`() {
         val lottoTickets = listOf(
@@ -26,7 +22,7 @@ internal class LottoScoreTest {
             LottoNumber(7)
         )
 
-        val lottoResults = LottoScore().compareNumber(winningLotto, LottoTickets(lottoTickets))
+        val lottoResults = LottoTickets(lottoTickets).compareNumber(winningLotto)
         val expected = mapOf(
             LottoPrize.FIRST to 1,
             LottoPrize.SECOND to 1,
@@ -44,26 +40,6 @@ internal class LottoScoreTest {
                 }
             }
         )
-    }
-
-    @ParameterizedTest
-    @CsvSource(
-        value = [
-            "5000, 12",
-            "60000, 1",
-            "60500, 1",
-            "120000, 0.5",
-            "210000, 0.285"
-        ]
-    )
-    internal fun `당첨 결과에 따라 수익률을 반환한다`(price: Int, expected: BigDecimal) {
-        // 총 상금: 60,000
-        val lottoResult = listOf(
-            LottoResult(LottoPrize.FOURTH, 1),
-            LottoResult(LottoPrize.FIFTH, 2)
-        ).toLottoResults()
-        val result = LottoScore().rateOfResult(LottoPrice(price), lottoResult)
-        assertThat(result).isEqualTo(expected.setScale(2, RoundingMode.HALF_UP))
     }
 
     private fun LottoTicket(vararg numbers: Int) =
